refactor sentence and para tokens in HtmlReader
Iterator interface of HtmlReader now generates a pair of tokens for
structural units: sentences and paragraphs. <s></s> and <p></p> are
used as token type strings for start/end tokens of each kind,
respectively. Sentence and paragraph text are available in a value of
the closing tokens (due to iterparse method design in
ElementTree). Opening tokens return None as a token value to make it
clear that sentence/paragraph text is not yet available.