DictReader: new options: canonical and glossfields
Options (1) to select a canonical lemma (an artical leading word from \le)
for any variant of that lemma (to be used in daba2vert and elsewhere
for lemmatization); (2) to provide the list and the order of the
fields to search for a gloss in an artical (earlier it was hardcoded).

These changes are used in daba2vert and utomm scripts.