switch from cElementTree to ElementTree for html in/out
This change is due to the apparent inability of cElementTree to
correctly process unicode combining diacritical marks in the element
attribute values (caused a parser error). Since with the introduction
of arbitrary attributes on tokens (e.g. free translations for sent)
more of these cases are expected, we had to trade off efficiency for
parsing reliability.

Accompanying change: provide XMLParser with encoding set to utf-8 to
make sure unicode in attribute values work correctly.