toolbox2vert: implement charwidth-based morpheme alignment
SIL Toolbox uses whitespace alignment to represent morphemic structure
of the token. As a pre-Unicode tool, it aligns using the number of
actual bytes used by the text that is variable for characters in
UTF-8. Additional config values ("alignment" block) is added to
explicitly point to a pair of fields (token/morpheme) that should be
used to align morphemes based on character width when converting from
Toolbox standard format.